Trait comp442_compiler::lexer::lexer::LexerAnalyzer

pub trait LexerAnalyzer {
    type TokenOutput;
    fn back(&mut self);

    fn back_n(&mut self, n: usize);

    fn forward(&mut self);

    fn forward_n(&mut self, n: usize);

    fn peek(&self) -> Option<char>;

    fn peek_n(&self, n: usize) -> Option<char>;

    fn next_char(&mut self) -> Option<char>;

    fn next_token(&mut self) -> Option<Self::TokenOutput>;

    fn skip_whitespace(&mut self);
}

Interface for a Lexer Analyzer

Associated Types

type TokenOutput

Loading content...
Required methods

fn back(&mut self)

moves the cursor back 1 character

fn back_n(&mut self, n: usize)

moves the cursor back n characters

fn forward(&mut self)

moves the cursor forwards 1 character

fn forward_n(&mut self, n: usize)

moves the cursor forwards n characters

fn peek(&self) -> Option<char>

Returns the next character in the input stream without advancing the lexer

fn peek_n(&self, n: usize) -> Option<char>

Returns the character n positions after the current position in the input stream without advancing the lexer

fn next_char(&mut self) -> Option<char>

Returns the next character, advancing the lexer

fn next_token(&mut self) -> Option<Self::TokenOutput>

Returns the next token, advancing the lexer

fn skip_whitespace(&mut self)

skips any whitespace at the beginning of the input

Loading content...