Enum comp442_compiler::parser::grammar::NamedSymbol

pub enum NamedSymbol {
    Start,
    AddOp,
    ArithExpr,
    ArraySize,
    ArraySizeAmb1,
    AssignOp,
    ClassDecl,
    Expr,
    ExprAmb1,
    Factor,
    FactorAmb1,
    FactorAmb2,
    FuncBody,
    FuncDecl,
    FuncDeclAmb1,
    FuncDef,
    FuncHead,
    FuncHeadAmb1,
    FuncHeadAmb2,
    FuncParams,
    Indice,
    MemberDecl,
    MultOp,
    OptClassDecl,
    OptFuncBody,
    Params,
    Prog,
    RelExpr,
    RelOp,
    ReptClassDecl,
    ReptFuncBody,
    ReptFuncParams0,
    ReptFuncParams1,
    ReptFuncParamsTail,
    ReptOptClassDecl,
    ReptOptFuncBody,
    ReptParams,
    ReptProg0,
    ReptProg1,
    ReptStatBlock,
    ReptVarDecl,
    ReptVariable,
    RightRecArithExpr,
    RightRecTerm,
    Sign,
    StatBlock,
    Statement,
    StatementAmb1,
    StatementAmb2,
    StatementAmb3,
    Term,
    Type,
    VarDecl,
    Variable,
    VariableAmb1,
    Visibility,
}

A named symbol is a non terminal in the grammar

Variants

Start

AddOp

ArithExpr

ArraySize

ArraySizeAmb1

AssignOp

ClassDecl

Expr

ExprAmb1

Factor

FactorAmb1

FactorAmb2

FuncBody

FuncDecl

FuncDeclAmb1

FuncDef

FuncHead

FuncHeadAmb1

FuncHeadAmb2

FuncParams

Indice

MemberDecl

MultOp

OptClassDecl

OptFuncBody

Params

Prog

RelExpr

RelOp

ReptClassDecl

ReptFuncBody

ReptFuncParams0

ReptFuncParams1

ReptFuncParamsTail

ReptOptClassDecl

ReptOptFuncBody

ReptParams

ReptProg0

ReptProg1

ReptStatBlock

ReptVarDecl

ReptVariable

RightRecArithExpr

RightRecTerm

Sign

StatBlock

Statement

StatementAmb1

StatementAmb2

StatementAmb3

Term

Type

VarDecl

Variable

VariableAmb1

Visibility

Implementations

impl NamedSymbol

pub fn first_set(&self) -> &[GrammarSymbol]

Returns the first set of the given non terminal

pub fn follow_set(&self) -> &[GrammarSymbol]

Returns the follow set of the given non terminal